Employee Leave Management System
The primary goal of the Employee Leave Management System Project is to handle the information of Employee, Leave, Company, Payroll, and Salary. It handles all employee information such as leave type, salary, and employee. Because the project is entirely constructed at the administrative level, only the administrator is assured access. The project's goal is to create application software that will decrease the amounts of manual labors required to manage Employee, Leave, Leave Type, and Company. It keeps a record of all the information regarding the company, such as payroll and salary.